MealTemplates tested in test_models.py

+++ meals/models.py
@@ -44,5 +44,5 @@ class MealTemplate(models.Model):
 
 	user = models.ForeignKey('auth.User',on_delete=models.CASCADE)
 	name = models.TextField(blank=False)	
-	cals = models.DecimalField(max_digits=4,decimal_places=2,blank=False)
+	cals = models.IntegerField(blank=False)
